# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CCE API",
    "description": "Code context engine: indexing, search, relations and tool endpoints",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/index/entity/{id}": {
      "delete": {
        "tags": [
          "Storage"
        ],
        "summary": "Handle delete entity request",
        "description": "Removes a specific entity and all its associated data\nacross all storage backends, scoped to the specified project.",
        "operationId": "handle_delete_entity",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "description": "Project ID for scoped storage operations",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Entity id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteEntityResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/project/{project_id}/relations/classification/stats": {
      "get": {
        "tags": [
          "Entity"
        ],
        "summary": "GET /api/project/{project_id}/relations/classification/stats",
        "description": "Returns counts of resolved relations grouped by external call classification.",
        "operationId": "get_classification_stats",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationStatsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Index unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/{project_id}/relations/classification/{classification}": {
      "get": {
        "tags": [
          "Entity"
        ],
        "summary": "GET /api/project/{project_id}/relations/classification/{classification}",
        "description": "Returns all resolved relations whose `external_type` matches the given\nclassification category (stdlib, external, dev, local, unknown).",
        "operationId": "get_relations_by_classification",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results (default 1000)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "project_id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "classification",
            "in": "path",
            "description": "Relation classification",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matched relations as untyped JSON array"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Index unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassificationStatsResponse": {
        "type": "object",
        "description": "Classification statistics response",
        "required": [
          "stats",
          "total"
        ],
        "properties": {
          "stats": {
            "type": "object",
            "description": "Per-classification counts",
            "additionalProperties": {
              "type": "integer",
              "minimum": 0
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of classified relations",
            "minimum": 0
          }
        }
      },
...
      "DeleteEntityResponse": {
        "type": "object",
        "description": "Delete entity response",
        "required": [
          "success",
          "message",
          "entity_id",
          "vectors_deleted",
          "bm25_documents_deleted",
          "relations_deleted",
          "elapsed_ms"
        ],
        "properties": {
          "bm25_documents_deleted": {
            "type": "integer",
            "minimum": 0
          },
          "elapsed_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "entity_id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "message": {
            "type": "string"
          },
          "relations_deleted": {
            "type": "integer",
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          },
          "vectors_deleted": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "description": "Error detail structure",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Common error response",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}